truck_api/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
PyJWT==2.9.0
pytz==2025.2
request2==0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
